requirements.txt
Flask==3.0.0
cs50==9.3.2
Flask-Session==0.5.0
Werkzeug==3.0.1
python-dotenv==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
